Fix bug #410: Incorrect video mode info shown when viewing HDMI IN live (T4) - second try
In [ServiceInfo] check whether the current service is HDMI IN and
return an empty string for VideoInfo if it is.

Return appropriate values for ServiceInfo data in other cases where
the current serviice is HDMI IN.

Also added idServiceHDMIIn serviceref type to eServiceReference
(lib/service/iservice.h).

Restore code in eServiceHDMI::getInfoString(int w) to return the
serviceref string that was introduced in commit bd0e4d0 and removed
when commit 7173655 reversed commit 1eb2f31.

Replaced some if statements with if expressions.