[serversocket] Set socket listen() backlog to a valid value
If backlog is set to zero, connections will get dropped if the thread is
doing anything else. This can result in erratic behaviour at the best of
times. With TCP SYN cookies enabled, this can lead to errors such as:

TCP: request_sock_TCP: Possible SYN flooding on port 8001. Dropping request.  Check SNMP counters.

A backlog of at least one is needed, however a larger value may prove to
be a better choice.