[InfoBarGenerics] Unnecessary instances of AutoTimer and AutoPoll created.
Rewrite the code for the actions on long-GREEN to use the plugin
descriptors to find the function to call rather than by creating
unnecessary new instances of the AutoTimer and AutoPoller.

The attempts in the original code to stop the AutoPoller while
editing fail because they only stop the new instance of AutoPoller,
not the one in the plugin.