eTextPara: Another bounding box fix
The bounding box calculation tests the bounding box of each glyph, but
they may have a slight offset.  When the box is used as the size (e.g.
to resize a MessageBox) this could result in the wrong size being used
(because the right bound is correct, but the left is offset, so the size
is off by the offset).  Remedy this by testing the x position of the
glyph, not the left bound.