[NetworSetup] Review
We link to have same screens in:
*FTP and TELNET .. use same code.
*We eliminate unnecessary screens in xml.
*We remove backgroundColor to follow the same skin default style.
*Added missing code..autoresize and WarningPixmap in messagebox.
=================================================== ======
@persianpros NetworkAFP  Service .....
In order to install and uninstall this service in openvision sources, you must add "netatalk" recipes-connectivity/netatalk to openvision-oe (the same as OE-A).
in the recipe we will prefer version 2.2.3 .. for this add in openvision-common.conf this:
PREFERRED_VERSION_netatalk? = "2.2.3"
and finally in openvision-enigma2-feed.bb add netatalk.