[config.py] Allow non string values in ConfigSelection
According to the comments in this file "all ids MUST be plain strings" but the reality is non-string values (i.e. boolean and integer) are found in many places (often plugins where the code is not under the umbrella of the distro it is being run on). This commit attempts to parse these non-string values gracefully so the offending code can still be run without problems.

Code from PRL